refactor(paging): split mechanism and policy, fix PL011 dedup, archive cleanup
- paging/descriptor.rs: unified VMSAv8-64 descriptor constants
- paging/walker.rs: walk_to_l3, split_block, TLBI, barriers (pure mechanism)
- paging/mapper.rs: PageTable with map/unmap/translate/map_region
- mmu.rs: policy only (PageTableBuilder, PageTableGuard, runtime ops)
- early_mmu.rs, mmu.rs: migrate to descriptor::* constants
- kernel/pmm.rs: add alignment assert in PmmBitmap::new()
- platform-qemu: replace local Pl011Uart with boot_common::pl011::Pl011
- target-qemu-armv7a: use boot_common::pl011::Pl011 with static lifetime
- workspace: remove 4 archived boot_legacy crates from members
- fix arithmetic overflow in dram_4gb() on 32-bit targets

+++ vivanta-boot/kernel/src/memory/resource.rs
@@ -76,7 +76,7 @@ pub struct MemoryProperties {
 impl MemoryProperties {
     pub const fn dram_4gb() -> Self {
         MemoryProperties {
-            capacity: 4 * 1024 * 1024 * 1024,
+            capacity: (4u64 * 1024 * 1024 * 1024) as usize,
             latency_ns: 80,
             bandwidth_mb_s: 25000,
             persistence: PersistenceType::Volatile,
